requirements.txt
langchain==0.0.166
faiss-cpu==1.7.4
openai==0.27.8
streamlit==1.23.1
python-dotenv==1.0.0
pypdf==3.9.1